changed unittest file, now testing in correct dnp_tools test
